Automatically proxy `ws://` URLs via Timing71 server.
Unfortunately, there's no way of allowing the plugin (running in an
HTTPS secure environment) to connect to an insecure WebSocket without
asking the user to change site security settings (yuck). We can use an
nginx proxy configured for the purpose; that obviously won't scale, but
it's only Natsoft that uses such URLs for the time being.

And, of course, this wouldn't show up in development because running on
`localhost` is special...