Guard against another sort of ghost stint.
If the timing screen flickers to blank (as the ACO one is wont to do)
then we don't want to create a new fake stint when we reappear. If we
were in the middle of a stint, then we should record the laptime as we
won't know if it's a new time or a duplicate.