Fix SocketIO websocket close handling.
We were previously defaulting to auto-reconnect, which means that when
the websocket server died we weren't falling back to polling as we
ought.

The previous `uuid` parameter was being ignored (the API for
`createWebsocket` changed but this call got missed) so we weren't using
the `tag` functionality - all this means is that we'll open up a new
websocket on page refreshes rather than trying to reuse any existing
ones. But since the host iframe disappears when the timing page reloads,
I'm not sure that's even a thing to care about any more.