Support analysis for legacy services.
We don't try to recreate the piecemeal state updates of the legacy
system with its multiple WAMP subscriptions; instead we get and migrate
the initial analysis state, then apply our own incremental updates when
we get a new service state.

We have a bunch of sticking plasters covering various bits of metadata
that aren't available in the legacy data (start time, last-updated
timestamp).